+++ tests/Integration/DataType/ProductAttributeTest.php
@@ -19,6 +19,7 @@ use OxidEsales\GraphQL\Storefront\Product\Infrastructure\Product as ProductInfra
 use OxidEsales\GraphQL\Storefront\Product\Service\Product as ProductService;
 use OxidEsales\GraphQL\Storefront\Product\Service\RelationService;
 use OxidEsales\GraphQL\Storefront\Shared\Infrastructure\Repository;
+use OxidEsales\GraphQL\Storefront\Shared\Infrastructure\ListConfiguration;
 use OxidEsales\GraphQL\Storefront\Shared\Service\Authorization;
 
 /**
@@ -88,7 +89,8 @@ final class ProductAttributeTest extends TestCase
     private function productRelationService(): RelationService
     {
         $repo = new Repository(
-            $this->createMock(QueryBuilderFactoryInterface::class)
+            $this->createMock(QueryBuilderFactoryInterface::class),
+            new ListConfiguration()
         );
 
         return new RelationService(