+++ tests/Integration/DataType/ProductImageGalleryTest.php
@@ -18,6 +18,7 @@ use OxidEsales\GraphQL\Storefront\Product\Infrastructure\Product as ProductInfra
 use OxidEsales\GraphQL\Storefront\Product\Service\Product as ProductService;
 use OxidEsales\GraphQL\Storefront\Product\Service\RelationService;
 use OxidEsales\GraphQL\Storefront\Shared\Infrastructure\Repository;
+use OxidEsales\GraphQL\Storefront\Shared\Infrastructure\ListConfiguration;
 use OxidEsales\GraphQL\Storefront\Shared\Service\Authorization;
 use PHPUnit\Framework\TestCase;
 
@@ -122,7 +123,8 @@ final class ProductImageGalleryTest extends TestCase
     private function productRelationService(): RelationService
     {
         $repo = new Repository(
-            $this->createMock(QueryBuilderFactoryInterface::class)
+            $this->createMock(QueryBuilderFactoryInterface::class),
+            new ListConfiguration()
         );
 
         return new RelationService(