+++ tests/Unit/DataType/ProductRelationServiceTest.php
@@ -17,6 +17,7 @@ use OxidEsales\GraphQL\Storefront\Product\DataType\Product;
 use OxidEsales\GraphQL\Storefront\Product\Infrastructure\Product as ProductInfrastructure;
 use OxidEsales\GraphQL\Storefront\Product\Service\Product as ProductService;
 use OxidEsales\GraphQL\Storefront\Product\Service\RelationService;
+use OxidEsales\GraphQL\Storefront\Shared\Infrastructure\ListConfiguration;
 use OxidEsales\GraphQL\Storefront\Shared\Infrastructure\Repository;
 use OxidEsales\GraphQL\Storefront\Shared\Service\Authorization;
 use PHPUnit\Framework\TestCase;
@@ -84,7 +85,8 @@ final class ProductRelationServiceTest extends TestCase
     private function productRelationService(): RelationService
     {
         $repo = new Repository(
-            $this->createMock(QueryBuilderFactoryInterface::class)
+            $this->createMock(QueryBuilderFactoryInterface::class),
+            new ListConfiguration()
         );
 
         return new RelationService(